+++ spec/spec_helper.rb
@@ -94,6 +94,8 @@ RSpec.configure do |config|
 =end
 
   config.before(:each) do
+    SSHKit.config.backend = SSHKit::Backend::Printer
+
     # block all system calls
     allow_any_instance_of(SSHKit::Interactive::Command).to receive(:system)