+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+### Change
+- run dummy command to force setting global netssh_options on host if using with capistrano
+
 ## 0.2.0 (2016-09-16)
 ### Feature
 - add DSL for interactive command invocation (`run_interactively`)